<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
v    If your PR doesn't close an issue, that's OK!  Just remove the Closes: #XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

Closes: #XXX

## What is the purpose of the change

[The benchmark ](https://github.com/osmosis-labs/osmosis/pull/5170) helped as see that having an additional cache layer in swaps would cause performance overhead.

This PR refactors swap out given in to remove an additional cache layer during swaps. It would still be present in calc (estimates) but not the core swaps logic. Calcs are not used by frontend or any other important clients so it is not as a big of a concern for launch.

Some graphs to show the improvement:

Old:

![image](https://github.com/osmosis-labs/osmosis/assets/34196718/f21c15b2-d83f-4ef4-8c98-30179052c0ac)

New:
![image](https://github.com/osmosis-labs/osmosis/assets/34196718/71a566df-6f97-4beb-b5f5-efe0e49e8c40)


Notice how the proportion of time spend in the iterator is noticeably smaller.

```
benchstat old.bench new.bench
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/osmosis/v15/x/concentrated-liquidity
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                     │  old.bench   │           new.bench           │
                     │    sec/op    │   sec/op     vs base          │
SwapExactAmountIn-16   777.9m ± 23%   750.3m ± 9%  ~ (p=0.436 n=10)
```

The p value here shows that the improvement isn't statistically significant but I believe that it is wrong since I see an improvement every time I run the test. Also, variance is decreased quite a bit.

Swap in given out will require the same changes in a separate PR.

## Testing and Verifying

This change is already covered by existing tests



## Documentation and Release Note

  - [ ] Does this pull request introduce a new feature or user-facing behavior changes?
  - [ ] Changelog entry added to `Unreleased` section of `CHANGELOG.md`?

Where is the change documented? 
  - [ ] Specification (`x/{module}/README.md`)
  - [ ] Osmosis documentation site
  - [ ] Code comments?
  - [x] N/A